Flow client: Implement fixed-time buffering (#97)
## Summary

This PR implements fixed-time buffering for incoming TTS audio. The
duration of the buffer is determined by the `audioBufferingMs` option.

This option already exists in the SDK, but was previously implemented
with a `setTimeout`, which did not guarantee a reliable buffer. This PR
changes the underlying logic to calculate based on the **time of the
audio itself** rather than delaying from the time the audio is received.

For example, if a buffer time of `500` is specified, the client will
buffer 500ms worth of audio before dispatching `"agentAudio"` events.
This makes the audio easier to understand in adverse network conditions.

**Note:** The queue is also automatically flushed whenever a
`ResponseCompleted` or `ResponseInterrupted` message is received. This
is to prevent audio getting stuck. This logic may need some finessing
but so far seems to be working well when the network is throttled.