Various fixes for Firefox (#102)
# Summary

This PR solves two separate problems encountered when using the Flow SDK
in Firefox:

## 1. Device enumeration

### Problem

On Firefox, unlike Chrome and Safari, the `enumerateDevices` method does
not return a list of media devices if not immediately preceded by the
creation of a `MediaStream` through `getUserMedia`.

This means that for users of the browser audio input package, granting
microphone permissions and then reloading the page fails to populate the
`deviceList`.

### Solution

When the permissions query comes back with a `"granted"` status, we take
a special action in the case of Firefox, creating a temporary
`MediaStream` before enumerating devices.

**Note**: There is one drawback to this solution: If the user accepts
microphone access, but does not click "Remember this decision" in the
popup, future navigations to the page will trigger an immediate
permissions popup. Given the constraints of Firefox, I'm not sure
there's anything we can do about this. They seem to be specifically
making it harder to arbitrarily list devices without user input
(probably to avoid fingerprinting)

## 2. Sample rate mismatch between recorded audio and TTS

### Problem

Due to [this bug](https://bugzilla.mozilla.org/show_bug.cgi?id=1725336),
Firefox can't capture audio at 16,000Hz (or any custom sample rate).

### Solution

Use a separate `AudioContext` for playback in Firefox, and allow it to
record at whatever sample rate it wants.