RT: Add option to disable timeout on `stopRecognition()` (#111)
Currently, the Node real-time example gives a timeout error when
streaming files longer than 10 seconds. This is because we call
`stopRecognition` when the data is finished streaming, but there is
still audio/transcription being processed before the server disconnects.

This PR allows the user to specify `noTimeout`, to disable waiting for
an `EndOfTranscript` message, facilitating faster-than-real-time use
cases.

Additionally cleaning up error messages and bumping versions of changed
packages.